requirements.txt
google-genai>=1.0.0
google-cloud-secret-manager>=2.20.0
google-cloud-storage>=2.14.0
supabase>=2.28.2
requests>=2.32.5
beautifulsoup4>=4.12.0
feedparser>=6.0.0
mlflow>=2.12.0
